Le Solveur enregistre mainteneant la fitnesse relative:
1=meilleur fitnesse (100% e corespondance)
0= pire fitnesse (0% de corespondance)
0.5 50% de  corespondance ext...

+++ Solve.h
@@ -65,12 +65,12 @@ class Solve{
   
     //Getters
     void affiche_final_fonction() const;
-    int* getHistoricFitness() const;
-    int getFinalFitness() const;
+    float* getHistoricFitness() const;
+    float getFinalFitness() const;
     
   protected:
     Fonction ** popFonct_ ; // pointeur d'un tableau de pointeurs
-    int * HistoricFitness_ ;
+    float * HistoricFitness_ ;
     int nbGeneration_;
     bool ** x_  ;
     bool * y_  ;
